ejudge-exporter/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
soupsieve==2.1
urllib3==1.26.2
PyYAML==5.4.1
click==8.0.1